reuse types
