+++ functions/common/types/index.ts
@@ -1,2 +1,3 @@
 export * from './environment';
 export * from './response';
+export * from './house';
