+++ functions/createHouse/src/createHouse.ts
@@ -2,9 +2,9 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { PutCommand } from '@aws-sdk/lib-dynamodb';
 
-import { Body } from "./types"
+import { House } from "../../common/types"
 
-export const createHouse = async (body: Body, tableName: string, client: DynamoDBClient) => {
+export const createHouse = async (body: House, tableName: string, client: DynamoDBClient) => {
   const putItemCommand = new PutCommand({
     TableName: tableName,
     Item: {
