+++ functions/createHouse/src/index.ts
@@ -1,6 +1,7 @@
 import { APIGatewayEvent, Handler } from 'aws-lambda';
 import { StatusCodes, ApiResponse, getEnv } from '../../common';
-import { Responsebody, Body } from "./types"
+import { Responsebody } from "./types"
+import { House } from "../../common/types"
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { DynamoDBDocumentClient } from '@aws-sdk/lib-dynamodb';
 
@@ -17,7 +18,7 @@ export const handler: Handler<APIGatewayEvent, ApiResponse<Responsebody>> = asyn
     return new ApiResponse(StatusCodes.error, { message: "EMPTY_BODY" })
   }
 
-  const body = JSON.parse(event.body) as Body
+  const body = JSON.parse(event.body) as House
   const result = await createHouse(body, tableName, ddb)
 
   const status = result ? StatusCodes.ok : StatusCodes.error;
