+++ functions/createHouse/src/types.ts
@@ -1,19 +1,3 @@
 export type Responsebody = {
   message: string;
 }
-
-interface Person {
-  name: string;
-  phone: string
-}
-
-export interface Body {
-  landlord: string;
-  houseId: string;
-  houseFriendlyName: string;
-  address: string;
-  details: string;
-  landlords: Person[],
-  leaseStatus: string
-  tenants: Person[]
-}
