+++ functions/getHouses/src/types.ts
@@ -1,3 +1,5 @@
+import {House} from "../../common/types/house"
+
 export interface Responsebody {
   houses: House[];
   message: string
@@ -5,21 +7,5 @@ export interface Responsebody {
 
 export interface Body {
   landlord: string
-}
-
-interface Person {
-  name: string;
-  phone: string
-}
-
-export interface House {
-  landlord: string;
-  houseId: string;
-  houseFriendlyName: string;
-  address: string;
-  details: string;
-  landlords: Person[],
-  leaseStatus: string
-  tenants: Person[]
 }
